+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 2.1.0 (2013-XX-XX)
 
+* Add the support of multiple managed files
 * Preserve other top-level keys than the configured one in the file
 * Add a rename map used to rename parameters when updating the parameters file
 * Add the possibility to use another top-level key than ``parameters``